feat: connect CLI to Personal Data Server  (#1)
* chore(release): 0.0.2

* feat: connect CLI to Personal Data Server

This PR ensures a compatible flow for the CLI by providing an option for people to bring their own Personal Data Server (PDS). If it isn't provided, we default to the bluesky server (https://bluesky.social) which is retrieved from the config.

I'm not certain of the correctness of that URL (https://bluesky.social), we can always update it later.

It also includes an account management module for overseeing actions on the provided PDS, such as creation and retrieval of posts for the purpose of testing. However, people can still decided to use the exposed methods &mdash; `createAccountOnPds`, `createPostOnPds`, getPostsFromPds` and `getPostCid` (if they want to reply a certain post) &mdash; in anyway they deem fit, should they provide their PDSes

* feat: add a new module to manage user session

- when you try to use the CLI for the first time after logging in, and try creating a test post, it'll throw an exception because the current session wasn't stored any where. The `atproto` API provides a method (`resumeSession`) to implement this.
- previously, we have prompts for the PDS ULR repeated quite a number of time during the auth flow. This commit addresses that by reading the config first befire prompting. We should only promt for a PDS url if it isn't available in the config
- used `credentialSession` instead of the soon-to-be-deprecated `AtpAgent` constructor alongside `Agent`