refactor: authentication and account management (#2)
Previously, the login command accepts a default `--pds` option with this value: https://bsky.app which in turn made `bsky-backup login` to fail whenever we try validating the PDS connection. Now, we're feeding this value from the user's config directly and if there's none, we fallback to a default again.

This refactor also allows users to login into multiple (bluesky/PDS) accounts at a time. Originally, the `accounts` list wasn't updated whenever we login, now, the handle and decentralized identifier of a user are stored together in the config so it is easier to logout of an account.

Retrieving posts from the PDS was futile at first because the `collection` value was hardcoded. Now, we obtain that by accessing the hostname of the provided PDS like so: `{new URL(config.pdsUrl).hostname}.feed.post`