feat: aadd BackupManager
`BackupManager` collocates all the supposed action of the CLI allowing people to backup their posts in one swoop with `bb backup posts`

The command fetches posts available in the pds (bluesky, eventually), save them locally, and offer to back it up on Storacha. If people happen to backup their posts and end up not uploading to Storacha immediately, they can still do it anytime, with this command: `bb backup upload`

This commit also introduces an update to the `PdsManager` command group that was formerly `test`. Now it is `admin`, we basically use it to carry out admin operations (create posts, list posts, signup etc) on the PDS.