feat: add StorachaAuth and backups manager (#3)
* chore(release): 0.0.2

* feat: add StorachaAuth and rename the bin to bb instead of bsky-backup

This module  exposes a couple of methods: `login` to authenticate with Storacha, `createStorachaSpace` to add a space for our uploads, and `listSpaces` does exactly what its name intends

* feat: aadd BackupManager

`BackupManager` collocates all the supposed action of the CLI allowing people to backup their posts in one swoop with `bb backup posts`

The command fetches posts available in the pds (bluesky, eventually), save them locally, and offer to back it up on Storacha. If people happen to backup their posts and end up not uploading to Storacha immediately, they can still do it anytime, with this command: `bb backup upload`

This commit also introduces an update to the `PdsManager` command group that was formerly `test`. Now it is `admin`, we basically use it to carry out admin operations (create posts, list posts, signup etc) on the PDS.

* fix: local backup of empty files

- if the posts length is not greater than zero, exit the prompt flow
- use a union of `Space` and `OwnedSpace` types for the `listSpaces` method
- get `ReadableStream` from blob
- fix logout method in `BlueskyAuth` not clearing the selected account

* fix: storacha upload to available space

Uploading to an available space always resulted in this error `Upload failed: failed space/blob/add invocation
` which is sorta related to permissions/proper delegations for the said space.

After taking a careful look at the `createSpace` method from the docs, i realized that it takes in a second parameter, `options`, that accepts the saved account. This led to an update of the `createStorachaSpace` that accepted just one arg previously to include the `account` arg needed by `createSpace`