feat: decode CAR data for JSON retrieval flow  (#5)
* chore(release): 0.0.2

* feat: decode CAR data for JSON retrieval flow

This PR introduces a refactor to the option that allows people to save their backups as JSON. The initial approach used the `litsRecord` method which is limited because we can't obtain the entire snapshot of the user's data due to pagination

Although, this method already returns JSON data for us, but it isn't exactly absolute. To solve this, we rely on the some packages from the IPLD ecosystem in support of DAG-CBOR ops which allows us to convert the stream of data from the `sync.getRepo` method.

This PR also included an update to the README that'll guide people on how to use the CLI.

More updates coming soon.

* fix list-posts, create-posts, replyTo posts

* chore: ignore files for npm