refactor: updated the storacha/bsky auth-to-upload flow
The major reason behind the issue we thought was fixed in #6 isn't really accurate. Sometimes, people do not authenticate with Storacha initially and they proceed to use the `bb backup posts` command which walks them through the process of selecting the data format (CAR|JSON), to authenticating with Storacha, choosing a space if they haven't and finally backing up their data.

Since this happens or at least, I experienced it a lot of time, we're always stuck on the 'Establishing connection..." step, because there's no email to auth with. The ideal thing to do is to prompt people to log in, if we can't find any auth session/state in the config. If there's an auth session, this flow works as it should.

This commit also includes typo fixes in the README and general updates to the 'how-to' guide